:recycle: Refactor: 페이징 방식에서 offset 방식으로 검색 방식 수정 (#66)
