+++ src/main/java/baeksaitong/sofp/domain/search/dto/KeywordDto.java
@@ -1,4 +1,4 @@
-package baeksaitong.sofp.domain.search.dto.request;
+package baeksaitong.sofp.domain.search.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
@@ -7,7 +7,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class KeywordReq {
+public class KeywordDto {
     @Schema(description = "검색 키워드 - 알약 이름, 성분, 효능 정보",example = "가스디알정")
     private String keyword;
     @Schema(description = "알약 모양", example = "원형")
@@ -20,8 +20,8 @@ public class KeywordReq {
     private String formulation;
     @Schema(description = "알약 분할선", example = "-/+")
     private String line;
-    @Schema(description = "검색 페이지 - 0부터 시작", defaultValue = "0")
-    private int page=0;
-    @Schema(description = "페이지 당 요소 갯수", defaultValue = "5")
-    private int limit=5;
+    @Schema(description = "마지막 요소(추가 요소 요청 시 이전 결과의 마지막 요소의 알약 ID", defaultValue = "0")
+    private Long lastId;
+    @Schema(description = "페이지 당 요소 갯수", defaultValue = "10")
+    private int limit=10;
 }
