+++ src/main/java/baeksaitong/sofp/domain/search/dto/response/KeywordDto.java
@@ -4,6 +4,8 @@ import baeksaitong.sofp.domain.pill.entity.Pill;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record KeywordDto(
+        @Schema(description = "알약 ID")
+        Long pillId,
         @Schema(description = "알약 시리얼 번호")
         Long serialNumber,
         @Schema(description = "알약 이름")
@@ -24,7 +26,8 @@ public record KeywordDto(
         Long FavoriteId
 ) {
         public KeywordDto(Pill pill, Boolean isWaring, Long favoriteId){
-                this(pill.getSerialNumber(),
+                this(pill.getId(),
+                        pill.getSerialNumber(),
                         pill.getName(),
                         pill.getClassification(),
                         pill.getProOrGen(),
